Update travis.yml to validate only on travis install
Travis runs the mvn install command, skipping tests in the travis install phase. This means the donut plugin fails as the integration tests have not been run at that stage. Changed the travis.yml to mvn initialize in the travis install phase, and to run the full tests in the script phase